refactor: intent key 통일
