+++ app/src/main/java/woowacourse/movie/seats/presenter/SeatsPresenter.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.seats.presenter
 
 import android.widget.TextView
 import woowacourse.movie.common.MovieDataSource
+import woowacourse.movie.common.MovieDataSource.movieList
 import woowacourse.movie.seats.contract.SeatsContract
 import woowacourse.movie.seats.model.Seat
 import woowacourse.movie.seats.model.SeatsDataSource
@@ -10,6 +11,7 @@ import woowacourse.movie.seats.model.SeatsDataSource.movieId
 import woowacourse.movie.seats.model.SeatsDataSource.seat
 import woowacourse.movie.seats.model.SeatsDataSource.seatTotalPrice
 import woowacourse.movie.seats.model.SeatsDataSource.selectedSeats
+import woowacourse.movie.seats.model.SeatsDataSource.theaterId
 import woowacourse.movie.seats.model.SeatsDataSource.ticketCount
 import woowacourse.movie.seats.model.SeatsDataSource.time
 
@@ -39,7 +41,8 @@ class SeatsPresenter(val view: SeatsContract.View) : SeatsContract.Presenter {
     override fun startNextActivity() {
         view.startNextActivity(
             movieId,
-            MovieDataSource.movieList[movieId.toInt()].title,
+            theaterId,
+            movieList[movieId.toInt()].title,
             date,
             time,
             selectedSeats,
@@ -89,5 +92,9 @@ class SeatsPresenter(val view: SeatsContract.View) : SeatsContract.Presenter {
     override fun clearSelectedSeats() {
         seatTotalPrice = 0
         Seat.seats.clear()
+    }
+
+    override fun storeTheaterId(theaterId: Long) {
+        SeatsDataSource.theaterId = theaterId
     }
 }
